+++ src/sgame/sg_bot_nav.cpp
@@ -116,7 +116,7 @@ float RadiusFromBounds2D( vec3_t mins, vec3_t maxs )
 {
 	float rad1s = Square( mins[0] ) + Square( mins[1] );
 	float rad2s = Square( maxs[0] ) + Square( maxs[1] );
-	return sqrt( std::max( rad1s, rad2s ) );
+	return sqrtf( std::max( rad1s, rad2s ) );
 }
 
 float BotGetGoalRadius( gentity_t *self )
@@ -512,8 +512,8 @@ bool BotFindSteerTarget( gentity_t *self, vec3_t dir )
 	for ( i = 0; i < 5; i++, yaw1 -= 15 , yaw2 += 15 )
 	{
 		//compute forward for right
-		forward[0] = cos( DEG2RAD( yaw1 ) );
-		forward[1] = sin( DEG2RAD( yaw1 ) );
+		forward[0] = cosf( DEG2RAD( yaw1 ) );
+		forward[1] = sinf( DEG2RAD( yaw1 ) );
 		//forward is already normalized
 		//try the right
 		VectorMA( self->s.origin, BOT_OBSTACLE_AVOID_RANGE, forward, testPoint1 );
@@ -530,8 +530,8 @@ bool BotFindSteerTarget( gentity_t *self, vec3_t dir )
 		}
 
 		//compute forward for left
-		forward[0] = cos( DEG2RAD( yaw2 ) );
-		forward[1] = sin( DEG2RAD( yaw2 ) );
+		forward[0] = cosf( DEG2RAD( yaw2 ) );
+		forward[1] = sinf( DEG2RAD( yaw2 ) );
 		//forward is already normalized
 		//try the left
 		VectorMA( self->s.origin, BOT_OBSTACLE_AVOID_RANGE, forward, testPoint2 );
